GPII-1167:  Split apart tests by category.  Fixed initial wiring for requestAware grade by adding a merge policy block.  Renamed "session aware" to "request aware" to better reflect its purpose.
