+++ index.js
@@ -27,5 +27,5 @@ require("./src/js/router");
 // A router module to handle static content.  Wraps the built-in `express.static()` module.
 require("./src/js/static");
 
-// A convenience grade on which "session aware" components are based
-require("./src/js/sessionAware");
+// A convenience grade on which "request aware" components are based
+require("./src/js/requestAware");
