--- tests/js/index.js
@@ -1,8 +1,0 @@
-"use strict";
-require("./test-caseholder");
-require("./test-middleware");
-require("./test-request-aware-router");
-require("./test-router");
-require("./test-router-cookiesetter");
-require("./test-router-params");
-require("./test-router-reqview");
