+++ tests/js/middleware/test-middleware-counter.js
